LPS-71622 Move test resources
