+++ dynamic-data-mapping-io/src/test/java/com/liferay/dynamic/data/mapping/io/internal/BaseDDMFormSerializerTestCase.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.io;
+package com.liferay.dynamic.data.mapping.io.internal;
 
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
