+++ dynamic-data-mapping-io/src/test/java/com/liferay/dynamic/data/mapping/io/internal/DDMFormJSONDeserializerTest.java
@@ -12,86 +12,62 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.io;
+package com.liferay.dynamic.data.mapping.io.internal;
 
 import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldRenderer;
 import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldType;
 import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeServicesTracker;
 import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldTypeSettings;
-import com.liferay.dynamic.data.mapping.io.internal.DDMFormJSONSerializerImpl;
+import com.liferay.dynamic.data.mapping.io.DDMFormJSONDeserializer;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
+import com.liferay.dynamic.data.mapping.model.DDMFormField;
+import com.liferay.dynamic.data.mapping.model.DDMFormFieldValidation;
 import com.liferay.dynamic.data.mapping.model.DDMFormRule;
 import com.liferay.dynamic.data.mapping.model.DDMFormSuccessPageSettings;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormFieldTypeSettingsTestUtil;
 import com.liferay.portal.json.JSONFactoryImpl;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.ReflectionUtil;
 
 import java.lang.reflect.Field;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Test;
 
 import org.mockito.Matchers;
 import org.mockito.Mock;
 import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 
-import org.skyscreamer.jsonassert.JSONAssert;
-
 /**
  * @author Marcellus Tavares
  */
-public class DDMFormJSONSerializerTest extends BaseDDMFormSerializerTestCase {
+public class DDMFormJSONDeserializerTest
+	extends BaseDDMFormDeserializerTestCase {
 
 	@Before
 	@Override
 	public void setUp() throws Exception {
 		super.setUp();
 
-		setUpDDMFormJSONSerializer();
+		setUpDDMFormJSONDeserializer();
 	}
 
-	@Test
-	public void testDDMFormSerialization() throws Exception {
-		String expectedJSON = read("ddm-form-json-serializer-test-data.json");
+	@Override
+	protected DDMForm deserialize(String serializedDDMForm)
+		throws PortalException {
 
-		DDMForm ddmForm = createDDMForm();
-
-		ddmForm.setDDMFormRules(createDDMFormRules());
-		ddmForm.setDDMFormSuccessPageSettings(
-			createDDMFormSuccessPageSettings());
-
-		String actualJSON = _ddmFormJSONSerializer.serialize(ddmForm);
-
-		JSONAssert.assertEquals(expectedJSON, actualJSON, false);
+		return _ddmFormJSONDeserializer.deserialize(serializedDDMForm);
 	}
 
-	protected List<DDMFormRule> createDDMFormRules() {
-		List<DDMFormRule> ddmFormRules = new ArrayList<>();
-
-		DDMFormRule ddmFormRule1 = new DDMFormRule(
-			"Condition 1", Arrays.asList("Action 1", "Action 2"));
-
-		ddmFormRules.add(ddmFormRule1);
-
-		DDMFormRule ddmFormRule2 = new DDMFormRule(
-			"Condition 2", Arrays.asList("Action 3"));
-
-		ddmFormRule2.setEnabled(false);
-
-		ddmFormRules.add(ddmFormRule2);
-
-		return ddmFormRules;
-	}
-
-	protected DDMFormSuccessPageSettings createDDMFormSuccessPageSettings() {
-		return new DDMFormSuccessPageSettings("Body Text", "Title Text", true);
+	@Override
+	protected String getDeserializerType() {
+		return "json";
 	}
 
 	protected DDMFormFieldTypeServicesTracker
@@ -136,23 +112,29 @@ public class DDMFormJSONSerializerTest extends BaseDDMFormSerializerTestCase {
 		return ddmFormFieldTypeServicesTracker;
 	}
 
-	protected void setUpDDMFormJSONSerializer() throws Exception {
+	@Override
+	protected String getTestFileExtension() {
+		return ".json";
+	}
+
+	protected void setUpDDMFormJSONDeserializer() throws Exception {
 
 		// DDM form field type services tracker
 
 		Field field = ReflectionUtil.getDeclaredField(
-			DDMFormJSONSerializerImpl.class,
+			DDMFormJSONDeserializerImpl.class,
 			"_ddmFormFieldTypeServicesTracker");
 
 		field.set(
-			_ddmFormJSONSerializer, getMockedDDMFormFieldTypeServicesTracker());
+			_ddmFormJSONDeserializer,
+			getMockedDDMFormFieldTypeServicesTracker());
 
 		// JSON factory
 
 		field = ReflectionUtil.getDeclaredField(
-			DDMFormJSONSerializerImpl.class, "_jsonFactory");
+			DDMFormJSONDeserializerImpl.class, "_jsonFactory");
 
-		field.set(_ddmFormJSONSerializer, new JSONFactoryImpl());
+		field.set(_ddmFormJSONDeserializer, new JSONFactoryImpl());
 	}
 
 	protected void setUpDefaultDDMFormFieldType() {
@@ -173,8 +155,62 @@ public class DDMFormJSONSerializerTest extends BaseDDMFormSerializerTestCase {
 		);
 	}
 
-	private final DDMFormJSONSerializer _ddmFormJSONSerializer =
-		new DDMFormJSONSerializerImpl();
+	@Override
+	protected void testBooleanDDMFormField(DDMFormField ddmFormField) {
+		super.testBooleanDDMFormField(ddmFormField);
+
+		DDMFormFieldValidation ddmFormFieldValidation =
+			ddmFormField.getDDMFormFieldValidation();
+
+		Assert.assertNotNull(ddmFormFieldValidation);
+		Assert.assertEquals(
+			"Boolean2282", ddmFormFieldValidation.getExpression());
+		Assert.assertEquals(
+			"You must check this box to continue.",
+			ddmFormFieldValidation.getErrorMessage());
+
+		Assert.assertEquals("true", ddmFormField.getVisibilityExpression());
+	}
+
+	@Override
+	protected void testDDMFormRules(List<DDMFormRule> ddmFormRules) {
+		Assert.assertEquals(ddmFormRules.toString(), 2, ddmFormRules.size());
+
+		DDMFormRule ddmFormRule1 = ddmFormRules.get(0);
+
+		Assert.assertEquals("Condition 1", ddmFormRule1.getCondition());
+		Assert.assertEquals(
+			Arrays.asList("Action 1", "Action 2"), ddmFormRule1.getActions());
+		Assert.assertTrue(ddmFormRule1.isEnabled());
+
+		DDMFormRule ddmFormRule2 = ddmFormRules.get(1);
+
+		Assert.assertEquals("Condition 2", ddmFormRule2.getCondition());
+		Assert.assertEquals(
+			Arrays.asList("Action 3"), ddmFormRule2.getActions());
+		Assert.assertFalse(ddmFormRule2.isEnabled());
+	}
+
+	@Override
+	protected void testDDMFormSuccessPageSettings(
+		DDMFormSuccessPageSettings ddmFormSuccessPageSettings) {
+
+		Assert.assertNotNull(ddmFormSuccessPageSettings);
+		Assert.assertEquals("Body Text", ddmFormSuccessPageSettings.getBody());
+		Assert.assertEquals(
+			"Title Text", ddmFormSuccessPageSettings.getTitle());
+		Assert.assertTrue(ddmFormSuccessPageSettings.isEnabled());
+	}
+
+	@Override
+	protected void testDecimalDDMFormField(DDMFormField ddmFormField) {
+		super.testDecimalDDMFormField(ddmFormField);
+
+		Assert.assertEquals("false", ddmFormField.getVisibilityExpression());
+	}
+
+	private final DDMFormJSONDeserializer _ddmFormJSONDeserializer =
+		new DDMFormJSONDeserializerImpl();
 
 	@Mock
 	private DDMFormFieldType _defaultDDMFormFieldType;
