+++ dynamic-data-mapping-io/src/test/resources/com/liferay/dynamic/data/mapping/io/internal/dependencies/ddm-form-field-type-settings-layout-serializer-test-data-default-column-size.json
