+++ dynamic-data-mapping-io/src/test/resources/com/liferay/dynamic/data/mapping/io/internal/dependencies/ddm-form-json-deserializer-test-data.json
