+++ dynamic-data-mapping-io/src/test/resources/com/liferay/dynamic/data/mapping/io/internal/dependencies/ddm-form-xsd-deserializer-test-data.xml
